feat: InputParser 생성
InputParser를 생성하여 InputView의 책임을 분리한다.
